move filter control to store
